+++ src/components/film-popup/comment.js
@@ -1,4 +1,4 @@
-import {createElement} from "../dom-util";
+import {createElement} from "../../util/dom-util";
 
 export default class Comment {
   constructor(comment) {
