+++ src/components/film/films-block.js
@@ -1,12 +1,12 @@
-import {createElement} from "../dom-util";
+import {createElement} from "../../util/dom-util";
 
-export default class SiteMenuContainer {
+export default class FilmsBlock {
   constructor() {
     this._element = null;
   }
 
   getTemplate() {
-    return `<nav class="main-navigation"></nav>`;
+    return `<section class="films"></section>`;
   }
 
   getElement() {
