+++ src/components/site-menu/site-menu-filter.js
@@ -1,4 +1,4 @@
-import {createElement} from "../dom-util";
+import {createElement} from "../../util/dom-util";
 
 export default class SiteMenuFilter {
   constructor(filter, count) {
