+++ src/main.js
@@ -1,20 +1,20 @@
-import ProfileUserComponent from "./components/profile-user";
-import SiteMenuContainerComponent from "./components/site-menu-container";
-import SiteMenuItemsComponent from "./components/site-menu-items";
-import SiteMenuItemComponent from "./components/site-menu-item";
-import SiteMenuFilterComponent from "./components/site-menu-filter";
-import SortContainerComponent from "./components/sort-container";
-import SortComponent from "./components/sort";
-import FilmsBlockComponent from "./components/films-block";
-import FilmsListComponent from "./components/films";
-import FilmCardComponent from "./components/film-card";
-import ShowMoreButtonComponent from "./components/load-more-button";
-import FilmDetailsComponent from "./components/film-details-popup";
-import FilmListExtraComponent from "./components/film-list-extra";
-import CommentComponent from "./components/comment";
+import ProfileUserComponent from "./components/profile/profile-user";
+import SiteMenuContainerComponent from "./components/site-menu/site-menu-container";
+import SiteMenuItemsComponent from "./components/site-menu/site-menu-items";
+import SiteMenuItemComponent from "./components/site-menu/site-menu-item";
+import SiteMenuFilterComponent from "./components/site-menu/site-menu-filter";
+import SortContainerComponent from "./components/sort/sort-container";
+import SortComponent from "./components/sort/sort";
+import FilmsBlockComponent from "./components/film/films-block";
+import FilmsListComponent from "./components/film/films";
+import FilmCardComponent from "./components/film/film-card";
+import ShowMoreButtonComponent from "./components/film/load-more-button";
+import FilmDetailsComponent from "./components/film-popup/film-details-popup";
+import FilmListExtraComponent from "./components/film/film-list-extra";
+import CommentComponent from "./components/film-popup/comment";
 import {generateFilms, FILM_LIST_TITLE} from "./mock/film";
 import {generationComments} from "./mock/comments";
-import {render, RenderPosition} from "./dom-util";
+import {render, RenderPosition} from "./util/dom-util";
 
 const FILM_COUNT = 18;
 const SHOWING_FILMS_COUNT_ON_START = 5;
